# Cache Class#getSimpleName and Class#getCanonicalName