This patch introduces cache for `Class#getSimpleName` and `Class#getCanonicalName`

Benchmarks:

1. Startup. I wrote simple Spring Application and measured with JMH how much time it needs to start. In fact I haven't noticed significant deviations:
```
        Benchmark            Mode  Cnt  Score   Error  Units
Before: DemoApplication.run  avgt    5  0.121 ± 0.003   s/op
After:  DemoApplication.run  avgt    5  0.119 ± 0.003   s/op
```

2. Throughput. I also measured with JMH how performant is this implementations. Benchmark includes two types of executions: normal one (single-threaded) and `...Concurrent` one - 10 threads in parallel.
```
        Benchmark                                                                      Mode  Cnt          Score            Error  Units

Before: OpenJ9ClassMetadataThroughputBenchmark.classCanonicalNameBenchmark            thrpt    5   29182175.748 ±    1537962.440  ops/s
After:  OpenJ9ClassMetadataThroughputBenchmark.classCanonicalNameBenchmark            thrpt    5  219930389.533 ±    2246706.612  ops/s

Before: OpenJ9ClassMetadataThroughputBenchmark.classCanonicalNameBenchmarkConcurrent  thrpt    5  123943856.531 ±     991646.387  ops/s
After:  OpenJ9ClassMetadataThroughputBenchmark.classCanonicalNameBenchmarkConcurrent  thrpt    5  656622074.401 ±   12697222.008  ops/s

Before: OpenJ9ClassMetadataThroughputBenchmark.classSimpleNameBenchmark               thrpt    5    9446333.682 ±     148938.440  ops/s
After:  OpenJ9ClassMetadataThroughputBenchmark.classSimpleNameBenchmark               thrpt    5  213360602.590 ±   93528750.650  ops/s

Before: OpenJ9ClassMetadataThroughputBenchmark.classSimpleNameBenchmarkConcurrent     thrpt    5   30667390.182 ±     100680.850  ops/s
After:  OpenJ9ClassMetadataThroughputBenchmark.classSimpleNameBenchmarkConcurrent     thrpt    5  649498541.172 ±   11960827.470  ops/s
```
This benchmark looks even better than the previous one (using `ReflectCache`)

Fixes #9963

Signed-off-by: Alexey Anufriev <contact@alexey-anufriev.com>