@DanHeidinga 2020-10-20
Jenkins test sanity zlinux jdk8,jdk11