@alexey-anufriev 2020-10-20
> Looks pretty reasonable to me. @alexey-anufriev Can you share the performance results you had gotten in this issue for posterity?

Sure, just updated the description of the PR.