@alexey-anufriev 2020-10-23
> In this case, we're using SoftReferences which makes the caching more complicated as we need to deal with the case where the SoftReference's `referent` has been cleared. Once it's cleared by the GC, I think this cache will always return `null` rather than the correct string.

@DanHeidinga, I took into consideration this remark and adjusted the implementation.