@alexey-anufriev 2020-10-26
I did some memory consumption measurements using _spring-pet-clinic_ application.
I have measured the memory used by the instances of `java.lang.Class` class.

The results are the following:

### OpenJ9 stock

#### Right after the App start
```
Class Name     | Objects | Shallow Heap | Retained Heap
--------------------------------------------------------
java.lang.Class|  13,142 |    2,312,992 |    11,610,632
--------------------------------------------------------
```

#### After some App usage
```
Class Name     | Objects | Shallow Heap | Retained Heap
--------------------------------------------------------
java.lang.Class|  14,591 |    2,568,016 |    13,619,248
--------------------------------------------------------
```
---

### OpenJ9 patched

#### Right after the App start
```
Class Name     | Objects | Shallow Heap | Retained Heap
--------------------------------------------------------
java.lang.Class|  13,149 |    2,314,224 |    11,742,512
--------------------------------------------------------
```

#### After some App usage
```
Class Name     | Objects | Shallow Heap | Retained Heap
--------------------------------------------------------
java.lang.Class|  14,465 |    2,545,752 |    13,029,824
--------------------------------------------------------
```