@DanHeidinga 2020-10-27
@pshipton do you have any further comments on these changes?  Any other benchmarks you want to see run before merging?