@pshipton 2020-10-27
Is the plan to move some of the other fields into the MetadataCache afterwards? Such as cachedDeclaringClass, cachedEnclosingClass?