@alexey-anufriev 2020-10-27
> Is the plan to move some of the other fields into the MetadataCache afterwards? Such as cachedDeclaringClass, cachedEnclosingClass?

Exactly, the plan was to move these fields as well: `cachedDeclaringClass, cachedEnclosingClass, cachedInterfaces`.
Would you guys suggest me to do this separately or I can do this within the same PR?