@DanHeidinga 2020-10-27
I'd suggest doing that in a separate PR as this is otherwise ready to go.