@DanHeidinga 2020-10-27
Jenkins test sanity zlinux jdk8